+++ src/main/java/com/braintreegateway/PaymentMethodNonceDetails.java
@@ -6,11 +6,15 @@ public class PaymentMethodNonceDetails {
     private String cardType;
     private String lastTwo;
     private String lastFour;
+    private String username;
+    private String venmoUserId;
 
     public PaymentMethodNonceDetails(NodeWrapper node) {
         cardType = node.findString("card-type");
         lastTwo = node.findString("last-two");
         lastFour = node.findString("last-four");
+        username = node.findString("username");
+        venmoUserId = node.findString("venmo-user-id");
     }
 
     public String getCardType() {
@@ -23,5 +27,13 @@ public class PaymentMethodNonceDetails {
 
     public String getLastFour() {
         return lastFour;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public String getVenmoUserId() {
+        return venmoUserId;
     }
 }